refactor to override Group.make_context()
