+++ nipoppy/cli/__init__.py
@@ -3,7 +3,6 @@
 import sys
 from contextlib import contextmanager
 
-import rich_click as click
 from pydantic_core import ValidationError
 
 from nipoppy.env import BUG_REPORT_URL, DISCORD_URL
@@ -13,8 +12,6 @@ from nipoppy.logger import get_logger
 logger = get_logger()
 
 
-# TODO once logger is extracted from the workflows, we could remove the `workflow`
-# parameter and use as a standalone context manager
 @contextmanager
 def exception_handler(workflow):
     """Handle exceptions raised during workflow execution."""
@@ -45,42 +42,3 @@ def exception_handler(workflow):
         )
     finally:
         sys.exit(workflow.return_code)
-
-
-class OrderedAliasedGroup(click.RichGroup):
-    """Group that lists commands in the order they were added and supports aliases."""
-
-    alias_map = {
-        "doughnut": "track-curation",
-        "run": "process",
-        "track": "track-processing",
-    }
-
-    def list_commands(self, ctx):
-        """List commands in the order they were added."""
-        return list(self.commands.keys())
-
-    def get_command(self, ctx, cmd_name):
-        """Handle aliases.
-
-        Given a context and a command name, this returns a Command object if it exists
-        or returns None.
-        """
-        # recognized command
-        command = click.Group.get_command(self, ctx, cmd_name)
-        if command is not None:
-            return command
-
-        # aliases (to be deprecated)
-        try:
-            new_cmd_name = self.alias_map[cmd_name]
-        except KeyError:
-            return None
-
-        logger.warning(
-            (
-                f"The '{cmd_name}' subcommand is deprecated and will cause an error "
-                f"in a future version. Use '{new_cmd_name}' instead."
-            ),
-        )
-        return click.Group.get_command(self, ctx, new_cmd_name)
